add delete unused image functionality

+++ first-app-v5/controllers/admin.js
@@ -1,3 +1,5 @@
+const fileHelper = require('../util/file');
+
 const { validationResult } = require('express-validator');
 
 const Product = require('../models/product');
@@ -142,7 +144,8 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDescription;
 	  if (image) {
-		  product.imageUrl = image.path;
+		fileHelper.deleteFile(product.imageUrl);  
+		product.imageUrl = image.path;
 	  }
       return product.save().then((result) => {
         console.log("Updated Product");
@@ -177,8 +180,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProducts = (req, res, next) => {
 	const prodId = req.body.productId;
-	Product.deleteOne({ _id: prodId, userId: req.user._id })
-	// .findByIdAndRemove(prodId)
+	Product.findById(prodId)
+	.then(product => {
+		if (!product) {
+			return next(new Error('Product not found.'));
+		}
+		fileHelper.deleteFile(product.imageUrl);
+		return Product.deleteOne({ _id: prodId, userId: req.user._id });
+	})
 	.then(result => {
 		console.log("Destroyed Product");
 		res.redirect('/admin/products');
